+++ db/schema.rb
@@ -16,7 +16,7 @@ ActiveRecord::Schema.define(:version => 20120914101957) do
   create_table "achievement_points", :force => true do |t|
     t.integer  "user_id",        :null => false
     t.integer  "achievement_id", :null => false
-    t.integer  "point",          :null => false
+    t.float    "point",          :null => false
     t.string   "description"
     t.datetime "created_at",     :null => false
     t.datetime "updated_at",     :null => false
